Light now works bit better.. just trust..
scene a bit fucky now. only one scene can be rendered at once because otherwise it would be using the same static resources. Thats becuase I moved the field that were in the scene into light and made them static. (it looked nicer idfk)

This might have been real dumb but it cleaned up the scene a bit apart from the big fuckoff function to bind the scene which switches in the geometry textures for the light programs.

Something else needs to be worked out for this but it works for now.

still no shadows. did more reaserach stuff and I kinda worked out my plan was dog. like stencil buffers are a little bit faster but theyre harder to do and cos u need to work out the volume, they dont work with trasparent object and a bunch of other shit.

So i figured I'd just do the normal shadow map option. gotta work out texture cubes and aswell but other than that it should be good.

still also need to do directional and spot lights. i figured I would only need 1 directional light cos its always going to be the sun. I can tehn also add on the ambient pass there.

added an abstract mesh class which I mostly did to put random meshes in that. so far its only got a screen mesh in it but i could def put more in.

fixed problem with the camera which caused the projection matrix not to update.