New Computer so I ofc managed to fuck up the git source. Its fixed now i think. but i lost some progress doing it. I am a colossal dumb ass.
Found a new problem with shader program where Im dumb and my logic is bad. if a texture unit is removed there can be textures after it which wont get used because that was my break condition.

Also founds that the spatial  functionality doesnt work w the light. fuck knows why. there are floating spheres in the sky because the world transform isnt moving them to the right place.

the shaders also needed to be updated to be set a value because of an update to OpenTK

Also Added an aligned transform which doesnt contain the rotation matrix. this was mostly for Lights.