The wall people can hear you scream.
its coming together I guess? I'm not fuckin with serialization, I'm yoinkin code for that.

I wanna work on the fun stuff agaaaain. ffs.

 I can add components, remove them and it'll automatically assign the archetype. if an archetype has no entities related to it it'll cease. like you after the wall people have got to you.

Just started the behaviour object which will be added to the entity context to add functionality to groups of different entities that fit the Filter criteria.