So the behaviours work I think??
the biggest problem i can see at the minute is that the arhcetype object that the behaviour contains will keep it alive after its been destroyed in the entity context -> go fix.

Also this will be exported to a seperate solution so i can actually make use of the internal stuff and keep it a bit cleaner.

The main difference is that you will know inherit behaviour function(which is now called typedbehaiour) instead of a delegate function. I just thought it made more sense.

Added a lil bit of debug stuff