Moved files because funny.
removed unnecessary casting for typed behaviour.
Changed how ComponentType ID is accessed. its now through Component Manager instead of the individual class.
Added a Destroy to archetype which removes is to remove all references to the archetype so that it can be cleaned up by the garbage collection.
Added a typed indexer to the component pool. and made non specific indexer only for when its a inteface.
Moved everything to new files because, as previously stated, continuity of naming and changes has and will never hold me back from doing what is sexy.