idek anymore.
Ive been trying to implement a hierarchal transform system. problems.

I added the graph traversal thing for archetypes. this means everytime it finds a step up or step down from the current archetype it stores that data in a array so next time its a look up instead of a search. the reason i didnt do this is currently its just stored as an array of 256 bytes which isnt too bad but its on every archetype and idk how well it will scale.

I also removed the Destroy action on the archetype. idk it just didnt feel right.

Ive added some exception most of which im not using.

I added a getAllComponents for an entity idk when this might be useful but i have it now

Behaviours had a bunch of changes. which is really rude of it. everytime i make a minute change it requires a painfully going through each override. I also added threading functionality. idk if itll work very well but its there?

component is now an abstract class? idk if i like that.

exposed Component Manager because behaviour construction requires the ID? maybe? it doesnt but its exposed now?

I added a bunch of events to the entity class as well as the scene graph stuff.

Added a clear Components()
which just moves it to the Empty Archetype. and calls removed Components on all the lost components.

Ignore program im trying to set up a hierarchal transform stuff using a System/Behaviour and the entity scene graph. its not working well. the archetypes arent being useful here either. I think transforms will be largely different from most other Systems/Behaviours because its dependant on the parent aswell as the component.
my current approach to this is to have a list sorted by layer(ie how many parents it has). I can then loop through the sorted list and updating and itll prioritise the entities at the top of the scene graph.
the only semi useful article i found on this called it a dirty bit propergation and used two Systems/Behaviours: one to loop through and identify the dirty bits and a second to change the transforms which could work? its not as efficient as a sorted list but doesnt require a seperate list which needs to be maintained.