REWORKS BABEE. ComponentManager? whats that. sounds bad and stinky.
I've abstrated alot of the functionality away from the archetype and also added the functionality of the component manager. this may turn out to have been a mistake.

this isnt finished the entities arent being moved properly and they arent calling the right things. I also want a faster copying of types because previously it did a search for each component. I loop through both archetypes Type lists when a entity is moved and it would be good if i could use that.

Filter now uses Types to match the archetypes use. I figured this wouldnt make a huge performance impact because its a int instead of a byte comparison and the CPU handles thing in 32bit at least. idk i might be chatting shit.

Small changes to Entity mostly for previous changes. a couple of things cos they look nicer.

I still have all those exception and im still not using any.