bad stupid dumb. extra dumb. Sprinkles of chump-fuckery. hahaha yhhhh your stupid. how abt undoing all that work you just did? yh good job, good job. well done. dumbass.
previous stuff done bad and stupid. so alot needed to be undone and redone. dumbass.

Archetype works worse now. it might be slightly faster and it uses more of the data it has its just looks uglier. I'll work on it. okay it looks slightly nicer.

Ive removed alot of behaviours implementation becuase it was distracting. another stellar decision.

And its completely untested.