nested formatting? why not ya know..
nested formatting for behaviour and Archetype under Entity.

this is mostly so Archetype has access to the Entity._poolIndex and then so the behaviour class didnt look so stupidly long. This makes the syntax makes slightly more sense beacuse its an Entity.Archetype and and Entity.Behaviour.

I've also fixed all the function parameters to be less icky in Archetype