ya did it again..
Ive replaced the multi-parameter search stuff with just some look up tables. yhhh I'm a moron. the searches were cool tho.

Ive also added some events to the archetype for when an Entity is moved into the archetype. This is for a planned sorted Behaviour which will need to update when an Entity moves into a Behaviour. alternatively I could store a point to the behaviour in the archetype. this is tempting because it would be easier to search for new behaviours from the archetype. Its a bit of a monster class right now tho so maybe not.

The components are got through reference now which is nice. it means the structs actually work.

I also axed behaviour down even more. but itll be built up again.

finished setting up SetComponents which moves the entities archetype entirely.

ComponentManager now has an IDs<T1, T2, T3, T4>() function which returns an array of IDs which is very nice and useful and yes.