Brrrrrrr.... remove stufffff bye bye Entity Context. bye bye nested classes. bye bye Behaviour functionality. I love teletubbies.
instead of the Entity Context I have static Lists on the Archetype and Behaviour. i know it might be annoying in the future but it looked nicer alright.

the nested classes were only ever being used to access 1 or 2 attributes. so I axed it.

I also moved pools back to their own file. mostly to stop archetype looking so big.
Half the reason its so large is because the three MoveEntity functions.