hehe restructure files.
the last 5 commits a have been a clusterfuck of messing up the git stuff locally. so ive basically just deleted it and cloned it again. try not to do that again. it wasnt fun.
I then reorganised the files inside and added a deprecated folder which at some point i hope to remove.