Suddenly issuing troops from a defensive position. because sortie. yh.. jokes.. I've been sorting stuff again.
Archetype pools are now split again into "components" and "entities". felt like it.

MoveEntity is now public and takes an entity and an archetype that meant _archetypes and _poolIndex had to be public but honestly thats a dumbass thing to get caught up on.

new Exception!!! EntityNotFound(). its a real banger.

Group got renamed to System because why not. (nothing much else had changed on it)

The sorting now works with IComparable<ComponentSet> so I'm not repeating functions. Archetype and System now inherit IComparable<ComponentSet>. There are 2 new functions for this called BinaryFirst() and BinaryLast() which find the first and last index of the sorted list. This is for the Archetype Search.
The System will have a set of requirements of all, any or none which will use the sorted archetypes to search for all applicable archetypes using the first and last to find the edges.

Another thing that might speed things up a bit is a Lock on new components. meaning no new components can be registered and which would decrease the time for the comparisons. I don't think i want to do that tho.