Searching for love likes long strings of meaningless of 1s and 0s representing components. is an all or none kinda guy.
i got nothing more. im rly not funny.

the Archetype search works. it takes a query which formats the 3 arrays into 2 255bit arrays. then the sorted archetypes are split into ranges for each bit starting with the most significant. this means. if the bit is in the search query, performs the logic to as to whether the range should be kept. each range is continuously split by the divide of where the next bit is until there are no ranges left.

it works? so yaay? I havent finished the any filter. because i had it working wrong logically.

it should satisfy all AND any AND none.
before i had it set up like (all AND none) OR any. Which was dumb and bad.

Oh yh new Exception just dropped. I'm really excited for this one. It has huge potential. EntityALreadyExists(). It's going places you can tell. This is to prevent user trying to move entity to the archetype it already belongs to.

System is back to being called Behaviour. because System is a namespace and its kinda hard not to use.