Linear Searches Babee!!!
so cool what an interesting way of solving your problem!!! a linear search? who knew that could work?

The Archetype search was inefficient because they were forced to test individual bits. Thats bad that. its not awful but its not very smart either. thats not even mentioning memory efficiency. the stack of ints search data structs it has to store gets huge. the search worked.
but archetypes it wasnt even the larger issue. the search ran when a new behaviour was created. that means the at run time when a new archetype is created its still just as inefficient.

so fuck it. axed that shit. the function still exists in this version but ill probably remove it right after this.

but techniquely this is finished?